+++ src/apiflask/exceptions.py
@@ -31,7 +31,8 @@ class HTTPError(Exception):
         status_code: int,
         message: t.Optional[str] = None,
         detail: t.Optional[t.Any] = None,
-        headers: t.Optional[t.Mapping[str, str]] = None
+        headers: t.Optional[t.Mapping[str, str]] = None,
+        extra_fields: t.Optional[dict] = None
     ) -> None:
         """Initialize the error response.
 
@@ -43,10 +44,16 @@ class HTTPError(Exception):
                 provide the addition information such as custom error code,
                 documentation URL, etc.
             headers: A dict of headers used in the error response.
+            extra_fields: A dict of additioinal fields (custom error information) that will
+                added to the error response body.
 
         *Version changed: 0.9.0*
 
         - Set `detail` and `headers` to empty dict if not set.
+
+        *Version changed: 0.10.0*
+
+        - Add `extra_fields` parameter to accept additional error information.
         """
         super().__init__()
         if status_code not in default_exceptions:
@@ -59,6 +66,7 @@ class HTTPError(Exception):
         self.headers = headers or {}
         self.message = get_reason_phrase(status_code, 'Unknown error') \
             if message is None else message
+        self.extra_fields = extra_fields or {}
 
 
 class _ValidationError(HTTPError):
@@ -71,7 +79,8 @@ def abort(
     status_code: int,
     message: t.Optional[str] = None,
     detail: t.Optional[t.Any] = None,
-    headers: t.Optional[t.Mapping[str, str]] = None
+    headers: t.Optional[t.Mapping[str, str]] = None,
+    extra_fields: t.Optional[dict] = None
 ) -> None:
     """A function to raise HTTPError exception.
 
@@ -103,9 +112,15 @@ def abort(
             provide the addition information such as custom error code,
             documentation URL, etc.
         headers: A dict of headers used in the error response.
+        extra_fields: A dict of additioinal fields (custom error information) that will
+            added to the error response body.
 
     *Version changed: 0.4.0*
 
     - Rename the function name from `abort_json` to `abort`.
+
+    *Version changed: 0.10.0*
+
+    - Add new parameter `extra_fields`.
     """
-    raise HTTPError(status_code, message, detail, headers)
+    raise HTTPError(status_code, message, detail, headers, extra_fields)