+++ CHANGES.md
@@ -6,10 +6,13 @@
 - Revoke the undocumented name changes on `validates` and `validates_schema` from marshmallow ([issue #62][issue_62]).
 - Only expose marshmallow `fields`, `validators`, and `Schema` in APIFlask.
 - Remove the `status_code` field from the default error response ([issue #124][issue_124]).
+- Add parameter `extra_fields` to `abort` and `HTTPError`, it accepts a dict that will be added
+to the error response ([issue #125][issue_125]).
 
 [issue_110]: https://github.com/greyli/apiflask/issues/110
 [issue_62]: https://github.com/greyli/apiflask/issues/62
 [issue_124]: https://github.com/greyli/apiflask/issues/124
+[issue_125]: https://github.com/greyli/apiflask/issues/125
 
 
 ## Version 0.9.0